Crate thread_token

Structs

Condvar

A Condition Variable

Mutex

A mutual exclusion primitive useful for protecting shared data

NonZeroUsize

An integer that is known not to equal zero.

Receiver

The receiving side of a channel.

Token

A Token is what TokenBucket::take_token returns. The token is automatically released when the Token is dropped. One can also explicitly release the token via Token::release, but this function does nothing but drop the token.

TokenBucket

A TokenBucket is a bucket containing a fixed number of “tokens”. Threads can take request to take out a token. If no tokens are available, the thread is blocked (while consuming no CPU time) until a token is released and returned to the bucket. When this happens, one of the threads waiting for a token is woken up and takes the token. There is no guarantee as to which thread in the queue gets woken up when a token is available.

Enums

TryRecvError

An error returned from the try_recv method.

Functions

unbounded

Creates a channel of unbounded capacity.